Split MPFA-O 3D file into multiple files
