+++ include/private/tdympfao3Dcoreimpl.h
@@ -1,5 +1,5 @@
-#if !defined(TDYMPFAO3DIMPL_H)
-#define TDYMPFAO3DIMPL_H
+#if !defined(TDYMPFAO3DCOREIMPL_H)
+#define TDYMPFAO3DCOREIMPL_H
 
 #include <petsc.h>
 
@@ -10,9 +10,6 @@ PETSC_EXTERN PetscErrorCode TDyMPFAOComputeSystem_BoundaryVertices_SharedWithInt
 PETSC_EXTERN PetscErrorCode TDyMPFAOComputeSystem_BoundaryVertices_NotSharedWithInternalVertices_3DMesh(TDy,Mat,Vec);
 PETSC_EXTERN PetscErrorCode TDyMPFAORecoverVelocity_3DMesh(TDy,Vec);
 PETSC_EXTERN PetscReal TDyMPFAOVelocityNorm_3DMesh(TDy);
-PETSC_EXTERN PetscErrorCode TDyMPFAORecoverVelocity_InternalVertices_3DMesh(TDy,Vec,PetscReal*,PetscInt*);
-PETSC_EXTERN PetscErrorCode TDyMPFAORecoverVelocity_BoundaryVertices_NotSharedWithInternalVertices_3DMesh(TDy,Vec,PetscReal*,PetscInt*);
-PETSC_EXTERN PetscErrorCode TDyMPFAORecoverVelocity_BoundaryVertices_SharedWithInternalVertices_3DMesh(TDy,Vec,PetscReal*,PetscInt*);
 PETSC_EXTERN PetscErrorCode TDyMPFAOIFunction_3DMesh(TS,PetscReal,Vec,Vec,Vec,void*);
 PETSC_EXTERN PetscErrorCode TDyMPFAOIJacobian_3DMesh(TS,PetscReal,Vec,Vec,PetscReal,Mat,Mat,void*);
 PETSC_EXTERN PetscErrorCode TDyMPFAOIFunction_DAE_3DMesh(TS,PetscReal,Vec,Vec,Vec,void*);
@@ -20,5 +17,5 @@ PETSC_EXTERN PetscErrorCode TDyMPFAOSNESFunction_3DMesh(SNES,Vec,Vec,void*);
 PETSC_EXTERN PetscErrorCode TDyMPFAOSNESJacobian_3DMesh(SNES,Vec,Mat,Mat,void*);
 PETSC_EXTERN PetscErrorCode TDyMPFAOSetFromOptions(TDy);
 PETSC_EXTERN PetscErrorCode TDyMPFAOSNESPreSolve_3DMesh(TDy);
-#endif
 
+#endif
