+++ src/tdycore.c
@@ -1,7 +1,7 @@
 #include <private/tdycoreimpl.h>
 #include <private/tdysaturationimpl.h>
 #include <private/tdypermeabilityimpl.h>
-#include <private/tdympfao3Dimpl.h>
+#include <private/tdympfao3Dcoreimpl.h>
 #include <private/tdympfaoimpl.h>
 #include <private/tdyeosimpl.h>
